% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_labelled.R
\name{read_labelled}
\alias{read_labelled}
\title{Импортирует данные из Excel файла с метками переменных и значений}
\usage{
read_labelled(file, data.sheet = 1, vars.sheet = 2)
}
\arguments{
\item{file}{Имя файла Excel с двумя листами: для данных и словаря}

\item{data.sheet}{Имя или номер листа книги, содержащего табличные данные}

\item{vars.sheet}{Имя или номер листа книги, содержащего словарь}
}
\description{
Импортирует данные из Excel файла с метками переменных и значений
}
\examples{
# read_labelled("example.xlsx")
# read_labelled("example.xlsx", "Data View", "Variable View")
}
