% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_labelled.R
\name{read_labelled}
\alias{read_labelled}
\title{Imports data frame from Excel file with variable labels}
\usage{
read_labelled(file, data.sheet = 1, vars.sheet = 2)
}
\arguments{
\item{file}{Excel workbook with two sheets}

\item{data.sheet}{Name or number of sheet that contains data}

\item{vars.sheet}{Name or number of sheet that contains data dictionary}
}
\description{
Imports data frame from Excel file with variable labels
}
\examples{
# read_labelled("example.xlsx")
# read_labelled("example.xlsx", "Data View", "Variable View")
}
